requirements.txt
# Core Flask Dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0

# Data Processing
pandas>=2.2.0
numpy>=2.1.0

# API & Web Requests
requests>=2.31.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0

# Database
SQLAlchemy==2.0.25

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Production Server
gunicorn==21.2.0
